server/requirements.txt
fastapi==0.111.0
httptools==0.6.1
ipykernel==6.29.5
matplotlib==3.9.1
openai==1.35.3
pip-chill==1.0.3
python-dotenv==1.0.1
sentence-transformers==3.0.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
